purple_agent/requirements.txt
# ═══════════════════════════════════════════════════════════════════
# Purple Agent -- Pinned Dependencies v7.0
# Python 3.11+
#
# All versions are pinned for reproducible builds. Do NOT use >= ranges.
# Upgrade only with full test suite validation.
# ═══════════════════════════════════════════════════════════════════

# Graph Engine
networkx==3.2.1

# Agentic Orchestration
langgraph==0.2.60
langchain-core==0.3.30

# LLM Client
openai==1.59.9

# HTTP Client (async)
httpx==0.27.2

# Protobuf
protobuf==5.29.3
grpcio-tools==1.69.0

# A2A Server
fastapi==0.115.6
uvicorn==0.34.0

# NLP -- spaCy NER
spacy==3.8.4

# Environment
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0

# Utilities
pydantic==2.10.5
